CVPN-939 Add TokioTunRaw - a platform independent tun RW module
The module is added to enable platforms other than linux to being able
to interact with tun. The module is based on tokio-tun, but simplified to
the specific use here.

The module supports async read and sync write, which are the only
behavior we need for Lightway to interact with Tun effectively.