CVPN-1818 Packet Accumulation Trait and its Factory for lightway-core
Interface for accumulating inside packets. The trait can be used to do some processing
to a stream of packets (e.g. FEC code)
store() adds the packet to the accumulator and returns AccumulatorState.
The AccumulatorState indicates whether the pkts in the accumulator is
ready to be retrieved.

When the pkts are ready to be retrieved, get_accumulated_pkts() returns the packets.