CVPN-1818 ToggleEncoding wire frame type
After lightway's state changes to Online, client may send a ToggleEncoding packet
to enable/disable the ingress packet accumulator on the server side.

When the server receives a ToggleEncoding packet that has a different setting, the
server will reply with the same ToggleEncoding packet. When the client receives it,
the client will also enable the ingress packet accumulator on the client side.

TODO: this way of enabling the packet accumulator is not reliable in a high packet loss
environment (as the ToggleEncoding packet could be lost). Is there a more reliable way?