CVPN-1818 Packet Accumulator Test for Connection in lw-core
Modified the unit test of Connection in lw-core:
If encoding (packet accumulator) is enabled,
1. Client sends a toggle encoding pkt to the server.
2. Server then replies the same toggle encoding pkt to the client
3. When client receives the pkt, a PacketEncoderToggled event will be emitted and
the first message will be triggered to be sent.

Note: NoOpPacketAccumulator is used in the test case. No encoding is actually happening
Main purpose of the test is to test the logic around supporting the use of the  packet accumulator.