client: Packet Codec
Packet Codec is supplied by lw-client to lw-core.

The packets in the encoder might have sit in the encoder for too long if inside packets are not frequently received. To avoid having a high packet delay, lw-server triggers a flush when the encoder is ready.

There could be stale states in the decoder. The decoder's stale state is cleaned periodically.