server: create ticker task on each connection
Lightway server on each client connection is trying to create a ticker task,
which can handle timers.
But since we previously used JoinSet and we dropped the set immediately
after spawning, the tasks were never scheduled.

To fix it, add a new api to spawn the task in tokio worker threads, so
that it will not be dropped.