core: encoding request pkt retransmit function
A callback can be provided by the client to schedule the next retransmission. Each re-transmission scheduled by a callback corresponds to a request via the request id. Hence, stale requests will not affect the latest requests and will be discarded accordingly.

The number of re-transmission is limited by a constant to avoid wasting network resources in case the server does not have encoding enabled.

The server only commits and give responses to the non-stale requests. An incoming request is not stale if and only if the request id is higher or equal to the id of the most recently acknowledged request.