client: Defer choosing best connection
A `defer_connect_timeout` field is added to ClientConfig. After going
online, we wait this duration before choosing the best connection. If
more than one connection goes online after this time, the final
connection chosen based on the order in the original `servers` array.

If `defer_connect_timeout` is 0, we simply take the first connection
that goes online.