client: Defer choosing preferred connection
A `preferred_connection_wait_interval` field is added to ClientConfig.
We wait this duration before choosing the best connection. If more than
one connection goes online after this time, the final connection chosen
based on the order in the original `servers` array.

If `preferred_connection_wait_interval` is 0, we simply take the first
connection that goes online.