client: monitor route updates for updating server route
We create a static server route for VPN server and 0/1 and 128/1 to
redirect all other traffic through tunnel.

Though this works in general, in UDP connections if client
floats (i.e move to different network) the server route still
points to the old network and this will break the connection.

So monitor route changes and update the server route if
there is a change in default route in the system.