client: only set keepalive timeout if not already pending
When try to send Keepalive immediately in Needed state,
commit e129a70faf6ac65151c38eb6196d621c4f32c415 introduces a crictial
bug where the timeout is reset after sending the Keepalive.
This is serious because the timeout value will never be triggered
if it is reset continuously. This makes keepalive timeout mechanism
fail to work during Tracer timeout or NetworkChange.

Commit 44efa8a637ce63ec0bcd6a971514b2afffb27263 tried to fix the bug, but
instead of fixing the timeout, it introduced another method using
failed_attempts counter to fix it. This is redundant since the existing
timeout mechanism is supposed to do this precisely.
Also it increases the timeout value to timeout * 3(MAX_FAILED_ATTEMPTS).

This commit does the following:
 - reverts the changes related to failed keepalive counter changes
 - Fixes resetting the timer using FusedFuture implementation
   FusedFuture enables the use of is_terminated() method on OptionFuture