client: keepalive: Trigger tracer detection once instead of spamming
When the client detects there is some network issue using tracer
timeout, it starts the keepalive to check connection health.
But since we do not reset the timeout, it keeps on triggering tracer
keepalive, which will reset the keepalive interval and make it not
send keepalive until timeout.

This commit fixes it by raising tracer trigger only once, based on
last_outside_data_received.